feat(app): update logos and branding to be Gitcoin-centered
- move <Head> from index page into _app to add favicon
- update onboard wallet connector logo and description to remove dpopp branding
- remove unused RINKEBY_RPC_URL

[#145]
